+++ gems/aws-sdk-core/CHANGELOG.md
@@ -1,6 +1,9 @@
 Unreleased Changes
 ------------------
 
+* Feature - Add `Seahorse::Util.host_label?` to check strings for valid RFC-3986 host labels.
+* Feature - Add `Aws::ARN#to_h`.
+
 3.108.0 (2020-09-25)
 ------------------